v0.8 Major Update with new detection features
TweakBlocker v0.8 adds a new check for illegal client-side packages that are not recognized by the server. This prevents spawning classes and loading tweaks on the fly after the client has already connected to the servers.

This version also has improved logging in Stealth Mode, cleaner code, and several important bug fixes.